requirements-travis.txt
Django
gunicorn
django-allauth
django-bootstrap4
psycopg2-binary
django-google-maps
Pillow